+++ src/components/App.js
@@ -1,9 +1,8 @@
 import React, { Component } from "react";
 import NavBar from "./NavBar.jsx";
-import GoalBox from "./GoalBox.jsx";
-import GoalBox2 from "./GoalBox2.jsx";
-import MediaCard from "./Cards.jsx";
 import Months from "./AllMonths.jsx";
+import June from "./jun.jsx";
+import June2 from "./jun2.jsx";
 
 //import { Switch, Route } from "react-router-dom";
 //import MyHabits from "./MyHabits/MyHabits.jsx";
@@ -15,6 +14,7 @@ class App extends Component {
         <NavBar></NavBar>
         <br />
         <Months></Months>
+        <June2></June2>
       </div>
     );
   }
